fixed problem negative value problem in pixy data and modified getWord to return positive value. code is now recognizing blocks sometimes but doesn't always match.

+++ NRGRobot2018/src/org/usfirst/frc948/NRGRobot2018/vision/IPixyLink.java
@@ -1,7 +1,7 @@
 package org.usfirst.frc948.NRGRobot2018.vision;
 
 public interface IPixyLink {
-	short getWord();
+	int getWord();
 	byte getByte();
 	void send(byte[] data); //i2c data send method
 }
